Move numba to its own file
